requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
grpcio==1.60.1
grpcio-tools==1.60.1
idna==3.6
importlib-resources==6.1.1
jsonpath-ng==1.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
openpyxl==3.1.2
overrides==7.7.0
pkgutil_resolve_name==1.3.10
ply==3.11
protobuf==4.25.2
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
robotframework==7.0
robotframework-assertion-engine==3.0.3
robotframework-browser==18.1.0
robotframework-excellib==2.0.1
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.3.0
robotframework-requests==0.9.6
rpds-py==0.18.0
typing_extensions==4.9.0
urllib3==2.2.0
wrapt==1.16.0
zipp==3.17.0
